feat(auth): usa snapshot RPC
